refactor: remove LeaveForm component and integrate its functionality into LeaveRequestView
fix: update useApi to improve error handling and loading state management

feat: add useToast composable for toast notifications

refactor: streamline agents store for better readability and performance

refactor: enhance leaves store with improved balance fetching and leave creation logic

refactor: optimize LeaveRequestView for better date handling and user feedback

+++ database/migrations/2026_05_08_095140_create_leaves_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('status')->default('pending');
             $table->date('date_start');
             $table->date('date_end');
-            $table->integer('working_days');
+            $table->decimal('working_days', 4, 1);
             $table->text('reason')->nullable();
             $table->text('refusal_reason')->nullable();
             $table->foreignId('approved_by')->nullable()->constrained('users')->nullOnDelete();
